+++ src/Exceptions/ProcessException.php
@@ -4,7 +4,7 @@ use Symfony\Component\Process\Process;
 
 interface ProcessException
 {
-    public function setProcess(Process $process);
+    public function setProcess(Process $process): void;
 
     public function getProcess(): Process;