+++ src/Exceptions/RuntimeException.php
@@ -4,9 +4,13 @@ use Symfony\Component\Process\Process;
 
 class RuntimeException extends \RuntimeException implements ProcessException
 {
+
+    /**
+     * @var Process
+     */
     protected $process;
 
-    public function setProcess(Process $process)
+    public function setProcess(Process $process): void
     {
         $this->process = $process;
     }