+++ tests/bootstrap.php
@@ -1,8 +1,11 @@
 <?php
 
-define('CRAFT_BASE_PATH', __DIR__);
-define('YII_DEBUG', true);
-define('TEST_FILE', '/tmp/craft-async-queue.txt');
+if (!defined('CRAFT_BASE_PATH')) {
+    define('CRAFT_BASE_PATH', __DIR__);
+    define('YII_DEBUG', true);
+    define('TEST_FILE', '/tmp/craft-async-queue.txt');
+}
+
 
 require_once __DIR__ . '/../vendor/autoload.php';
 require_once __DIR__.'/../vendor/yiisoft/yii2/Yii.php';